[modify] modify about define  .

+++ src/ocsfs_serv/src/ocsfs_client_handler.cpp
@@ -136,7 +136,7 @@ bool OCSFS_Client_Handler::send_data(const QString& src_client_id, const QString
 void OCSFS_Client_Handler::recv_data() {
     QByteArray recv_data  = this->socket->readAll();
 
-    if (recv_data.size() < OCSFS_PROTO_HEAD_LEN) {
+    if (recv_data.size() < (long long int)OCSFS_PROTO_HEAD_LEN) {
         this->step = 0;
         this->send_data(QString(OCSFS_CLIENT_ID_LEN, '0'), QString("error"));
         qDebug() << "data len error: " << recv_data.size();
